trying to finish it

+++ blog/views.py
@@ -1,10 +1,15 @@
-from django.shortcuts import render
 from django.utils import timezone
 from .models import Post, Comment
 from django.shortcuts import render, get_object_or_404
-from .forms import PostForm, CommentForm
+from .forms import PostForm, CommentForm, RegistrationForm
 from django.shortcuts import redirect
 from django.contrib.auth.decorators import login_required
+from django.template import RequestContext
+from django.shortcuts import render_to_response
+from django.contrib.auth.models import User
+from django.http import HttpResponseRedirect
+from rest_framework import generics
+from .serializers import CommentSerializer
 
 
 def post_list(request):
@@ -32,17 +37,18 @@ def post_new(request):
 
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == "POST":
-        form = PostForm(request.POST, instance=post)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            '''post.published_date = timezone.now()'''
-            post.save()
-            return redirect('post_detail', pk=post.pk)
-    else:
-        form = PostForm(instance=post)
-    return render(request, 'blog/post_edit.html', {'form': form})
+    if post.author==request.user :
+        if request.method == "POST":
+            form = PostForm(request.POST, instance=post)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                post.published_date = timezone.now()
+                post.save()
+                return redirect('post_detail', pk=post.pk)
+        else:
+            form = PostForm(instance=post)
+        return render(request, 'blog/post_edit.html', {'form': form})
 @login_required
 
 def post_draft_list(request):
@@ -54,12 +60,13 @@ def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
     post.publish()
     return redirect('post_detail', pk=pk)
-@login_required
 
+@login_required
 def post_remove(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect('post_list')
+    if post.author==request.user :
+        post.delete()
+        return redirect('post_list')
 
 def add_comment_to_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
@@ -86,3 +93,29 @@ def comment_remove(request, pk):
     post_pk = comment.post.pk
     comment.delete()
     return redirect('post_detail', pk=post_pk)
+
+def register_page(request):
+    if request.method == 'POST':
+        form = RegistrationForm(request.POST)
+        if form.is_valid():
+            user = User.objects.create_user(username=form.cleaned_data['username'],password=form.cleaned_data['password1'],email=form.cleaned_data['email'])
+            return HttpResponseRedirect('/')
+    form = RegistrationForm()
+    variables = RequestContext(request, {'form': form})
+    return render_to_response('registration/register.html',variables)
+
+def can_edit (request, pk):
+    post = get_object_or_404(Post, pk=pk)
+    post.author = request.user
+    if post.author==request.user:
+        return render(request, 'blog/post_detail.html', {'post':post, 'can_edit': True})
+    else:
+        return render(request, 'blog/post_detail.html', {'post':post, 'can_edit': False})
+
+class PostsList(generics.ListCreateAPIView):
+    queryset = Comment.objects.all()
+    serializer_class = CommentSerializer
+
+class PostsDetail(generics.RetrieveUpdateDestroyAPIView):
+    queryset = Comment.objects.all()
+    serializer_class = CommentSerializer
